Removed a static global variable from DataPoint prohibiting multiple data loading.
The DataPoint class had the static field `featureCount` which was
set to the number of features in the data used for the current task.
This field was queried by the normalizers and the LinearRegRank ranker
in order to determine array size required for their computations.

However, this system was only safe as long for the lifetime of the
JVM only one dataset was loaded. Multithreading was also a major issue.
Adding the SparseDataPoint test
broke the following tests because the field was in the wrong state.
This was even true when using the original constructor of
SparseDataPoint which used text input and had nothing to do with
the newly introduced constructor in the last commit.

Now, the number of features are derived from the concrete data at
hand. This requires an iteration through all data points.
All tests running.